pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "diagnostic-copilot"
version = "0.1.0"
description = "AI-powered rare disease diagnostic copilot"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2",
    "langchain-anthropic>=0.3",
    "anthropic>=0.40",
    "chainlit>=1.3",
    "pronto>=2.5",
    "pandas>=2.1",
    "pymongo>=4.6",
    "redis>=5.0",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "rapidfuzz>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
include = ["core*", "tools*", "agent*", "eval*", "scripts*"]
